python-developer/08-fastapi/szkolenie-04-autentykacja/requirements.txt
fastapi[standard]
passlib[bcrypt]
PyJWT
python-multipart
